requirements.txt
asgiref==3.5.2
autopep8==1.7.0
distlib==0.3.5
Django==4.1
djangorestframework==3.13.1
filelock==3.8.0
flake8==5.0.4
mccabe==0.7.0
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
virtualenv==20.16.3
psycopg2==2.9.3